rename and add all other new classes
