+++ hydromt/data_source/rasterdataset.py
@@ -17,8 +17,8 @@ from pystac import MediaType
 from hydromt._typing import Bbox, ErrorHandleMethod, Geom, TimeRange, TotalBounds
 from hydromt.data_adapter.rasterdataset import RasterDatasetAdapter
 from hydromt.data_source.data_source import DataSource
-from hydromt.drivers.rasterdataset_driver import RasterDatasetDriver
-from hydromt.drivers.zarr_driver import ZarrDriver
+from hydromt.driver.rasterdataset_driver import RasterDatasetDriver
+from hydromt.driver.zarr_driver import ZarrDriver
 
 _KNOWN_DRIVERS: Dict[str, RasterDatasetDriver] = {"zarr": ZarrDriver}
 logger: Logger = getLogger(__name__)
