+++ tests/drivers/test_pyogrio_driver.py
@@ -8,8 +8,8 @@ from pyogrio.errors import DataSourceError
 from shapely import box
 
 from hydromt._typing import Bbox
-from hydromt.drivers.pyogrio_driver import PyogrioDriver
-from hydromt.metadata_resolvers.convention_resolver import ConventionResolver
+from hydromt.driver.pyogrio_driver import PyogrioDriver
+from hydromt.metadata_resolver.convention_resolver import ConventionResolver
 
 
 class TestPyogrioDriver:
