+++ tests/drivers/test_zarr_driver.py
@@ -5,8 +5,8 @@ import numpy as np
 import pytest
 import zarr
 
-from hydromt.drivers.zarr_driver import ZarrDriver
-from hydromt.metadata_resolvers.convention_resolver import ConventionResolver
+from hydromt.driver.zarr_driver import ZarrDriver
+from hydromt.metadata_resolver.convention_resolver import ConventionResolver
 
 
 class TestZarrDriver:
